+++ src/client/java/de/selectiverender/SectionVisibility.java
@@ -1,0 +1,7 @@
+package de.selectiverender;
+
+public enum SectionVisibility {
+    UNCHANGED,
+    PARTIAL,
+    HIDDEN
+}